requirements.txt
fastapi==0.111.1
uvicorn==0.30.1
pandas==2.2.2
python-docx==1.1.2
PyPDF2==3.0.1
tiktoken==0.7.0
python-dotenv==1.0.1